First Check-in
Task 1: Be creative and put your own responsive design on the clockwork.web project.

Task 2: As you use the application, one of its core features is that it saves each API request to the database. We would like you to add a feature that shows all previous database entries when the page loads and when a new response from the API is returned.

Task 3: Give the user the ability to request a timezone and return the server time in that timezone. The API should also record the timezone that the user requested in the database.